% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\docType{package}
\name{banneR}
\alias{banneR}
\alias{banneR-package}
\title{`banneR` mimics UNIX(TM) original banner program in R.}
\description{
`banneR` provides an R `banner()`function that behaves similar to
UNIX(TM) `banner` program.
}
\author{
José R. Valverde, 2018
}
\keyword{internal}
