% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{em_network_bg}
\alias{em_network_bg}
\title{Expectation-Maximization (EM) for hot zones.}
\usage{
em_network_bg(
  y,
  X,
  X_bg,
  X_hz = X,
  L,
  lambda,
  lambda_hz,
  p_bg_init,
  control = list()
)
}
\arguments{
\item{y}{response vector}

\item{X}{expanded design matrix for hot zone counts, including intercept (see \code{expand_design})}

\item{X_bg}{design matrix for background counts, including intercept}

\item{X_hz}{expanded design matrix for hot zone indicators, including intercept (see \code{expand_design}); defaults to \code{X}}

\item{L}{Laplacian matrix}

\item{lambda}{prior precision scale for hot zone counts (or regularization penalty)}

\item{lambda_hz}{prior precision scale for hot zone indicators (or regularization penalty)}

\item{p_bg_init}{initial background probabilities, assumed binary}

\item{control}{list specifying running parameters (see \code{glm.control})}
}
\value{
list containing coefficients \code{coef}, \code{coef_bg}, and \code{coef_hz} for the count regressions on hot zones and background and logistic regression on hot zone indicators, respectively; estimated \code{P(Z_i = 1 | y)} as E-weights \code{p_bg}; and \code{deviance} and press-loop cross-validation loss \code{pl}.
}
\description{
Given the priors \code{theta ~ N(0, (lambda Omega)^-)} and
\code{omega ~ N(0, (lambda_hz Omega_hz)^-)}, with
\code{Omega = direct-sum_i t(X_i) * L * X_i} and
\code{Omega_hz = direct-sum_i t(X_hz_i) * L * X_hz_i},
Poisson regresses \code{y | Z = 1 ~ X} and \code{y | Z = 0 ~ X_bg} with
\code{Z ~ X_hz}, the "hot zone" indicators, also network regularized via
\code{omega}. Here \code{X} and \code{X_hz} are network-expanded designs
(see \code{expand_design}). The conditional regressions are performed via
conditional EM with hot zone indicators as latent.
}
