% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{em_network_rank}
\alias{em_network_rank}
\title{Expectation-Maximization (EM) for network basis rank determination.}
\usage{
em_network_rank(
  y,
  X,
  L,
  lambda,
  log_tau_p,
  v0 = 0,
  fam = poisson(),
  offset = 0,
  start = NULL,
  control = list()
)
}
\arguments{
\item{y}{response vector}

\item{X}{expanded design matrix, including intercept (see \code{expand_design})}

\item{L}{Laplacian matrix}

\item{lambda}{prior precision scale (or regularization penalty)}

\item{log_tau_p}{log prior for \code{tau} (see \code{tau_prior})}

\item{v0}{prior variance for non-selected basis, defaults to zero}

\item{fam}{exponential family distribution, such as \code{family} in \code{glm}; defaults to \code{poisson}}

\item{offset}{linear predictor offset, defaults to zero}

\item{start}{initial coefficient estimates}

\item{control}{list specifying running parameters (see \code{glm.control})}
}
\value{
list containing coefficients \code{coef}, estimated log posterior \code{log P(tau | theta, y)} in \code{logptau} as (log) E-weights, and \code{deviance}
}
\description{
Given the rank-selection prior \code{theta | tau ~ N(0, (lambda Omega)^-)},
with \code{Omega = direct-sum_i M_[tau[i]] * t(X_i) * L * X_i M_[tau[i]]},
regresses \code{y ~ X}, \code{X} an expanded design, under family \code{fam}
via conditional EM with ranks \code{tau} as a latent. \code{tau} has log
prior distribution specified in \code{log_tau_p}, defined with respect to
the maximum rank of \code{X}. FSAM is used to initialize the procedure.
}
