% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{expand_design}
\alias{expand_design}
\title{Network-expanded design.}
\usage{
expand_design(X, Phi, rank = rep(ncol(Phi), ncol(X)))
}
\arguments{
\item{X}{base design matrix, including intercept}

\item{Phi}{Laplacian eigenvectors}

\item{rank}{expansion ranks, a vector with the same length as \code{ncol(X)}; defaults to a full expansion with \code{rep(ncol(Phi), ncol(X))}}
}
\value{
the expanded design in a matrix with \code{rank} as attribute
}
\description{
Given a base design matrix \code{X} and the Laplacian eigenvectors
\code{Phi}, computes the expanded design
\code{[Diag(X_i) * Phi_[1:rank[i]]]_i} for the i-th predictor in \code{X}.
}
