% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{fsam_network}
\alias{fsam_network}
\title{Forward Stagewise Additive Modeling (FSAM).}
\usage{
fsam_network(
  y,
  X,
  L,
  lambda,
  fam = poisson(),
  maxit = max(attr(X, "rank")),
  trace = FALSE
)
}
\arguments{
\item{y}{response vector}

\item{X}{expanded design matrix, including intercept (see \code{expand_design})}

\item{L}{Laplacian matrix}

\item{lambda}{prior precision scale (or regularization penalty)}

\item{fam}{exponential family distribution, such as \code{family} in \code{glm}; defaults to \code{poisson}}

\item{maxit}{maximum number of iterations, defaults to maximum rank in \code{X}}

\item{trace}{print iteration progress messages?}
}
\value{
list containing coefficients \code{theta} and \code{deviance}
}
\description{
Performs the FSAM regression \code{y ~ X} under family \code{fam}, with
coefficients having prior distribution \code{N(0, (lambda t(X) * L * X)^-)},
with \code{L} taken as a differential operator such as a Laplacian.
}
