% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{get_index_rank}
\alias{get_index_rank}
\title{Sub-index rank selection.}
\usage{
get_index_rank(rank, tau)
}
\arguments{
\item{rank}{full rank indices, as returned from an expanded design \code{X}
with \code{attr(X, "rank")} (see also \code{expand_design})}

\item{tau}{rank subsets, with same length as \code{rank}}
}
\value{
index range
}
\description{
Gets the index range from \code{rank} based on the rank selection specified
in \code{tau}.
}
