% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{laplacian_marginal}
\alias{laplacian_marginal}
\title{Laplacian marginalization}
\usage{
laplacian_marginal(graph, v, weights)
}
\arguments{
\item{graph}{igraph object}

\item{v}{vertex set to be marginalized}

\item{weights}{edge weights}
}
\value{
Laplacian matrix.
}
\description{
Given a set \code{v} of vertices to be marginalized in a weighted
\code{graph} with edge \code{weights}, compute the Laplacian of
\code{G[-v]}.
}
