% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{make_fitter}
\alias{make_fitter}
\title{Create network regularized fitter.}
\usage{
make_fitter(lambda, X, L)
}
\arguments{
\item{lambda}{precision scale or regularization penalty}

\item{X}{design matrix}

\item{L}{Laplacian matrix}
}
\value{
fitter object from package \code{bsglm}
}
\description{
Create a regularized fitter based on a normal prior with zero mean and
precision \code{lambda * t(X) * L * X}.
}
