% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{plot_em_fit}
\alias{plot_em_fit}
\title{Plot EM hot-zone network fit.}
\usage{
plot_em_fit(y, X, X_bg, res)
}
\arguments{
\item{y}{response vector}

\item{X}{expanded design matrix for hot zone counts, including intercept (see \code{expand_design})}

\item{X_bg}{design matrix for background counts, including intercept}

\item{res}{list as returned from \code{em_network_bg}}
}
\value{
list with \code{fitted} values and \code{residuals}, as side effect (invisible)
}
\description{
Given the output of a EM hot-zone network fit, plots observed vs espected counts, estimated probabilities of background vs observed counts, and residuals vs expected counts.
}
