% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{press_loop}
\alias{press_loop}
\title{Leave-One-Out-Proxy (LOOP) for PRESS score}
\usage{
press_loop(fit, X = model.matrix(fit))
}
\arguments{
\item{fit}{model fit object}

\item{X}{design matrix to use in PRESS evaluation}
}
\value{
PRESS score
}
\description{
Given a model \code{fit}, compute PRESS LOOP score.
}
