% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{similarity_phi}
\alias{similarity_phi}
\title{Find hyper-prior network range parameter based on network distances.}
\usage{
similarity_phi(dist, median_similarity, interval = c(0, 1), ...)
}
\arguments{
\item{dist}{distance matrix}

\item{median_similarity}{median similarity distance}

\item{interval}{root search interval, default (0, 1)}
}
\value{
The hyper-prior network range parameter \code{phi}
}
\description{
For the similarity score defined as \code{S = exp(-dist / phi)}, find
\code{phi} such that the median similarity score computed from distances
\code{dist} is \code{median_similarity}.
}
