% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{tau_prior_rho}
\alias{tau_prior_rho}
\title{Prior for basis indicator \code{tau}, using hyper-parameter \code{rho}.}
\usage{
tau_prior_rho(K, rho, alpha_0, alpha_1)
}
\arguments{
\item{K}{maximum basis expansion}

\item{rho}{probability ratio \code{P(tau = k + 1 | tau > 1) / P(tau = k | tau > 1)}}

\item{alpha_0}{probability of rank > 0}
}
\value{
alpha_1 probability of rank > 1 given rank > 0

vector with probabilities \code{P(tau | K)}
}
\description{
Computes the prior rank probabilities \code{P(tau | K)} where \code{K} is
the maximum basis expansion based on parameters \code{alpha_0 = P(tau > 0)},
\code{alpha_1 = P(tau > 1 | tau > 0)}, and \code{rho}, where
\code{P(tau | tau > 1)} is proportional to \code{rho^{tau - 1}}.
}
