% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{banner}
\alias{banner}
\title{banner}
\usage{
banner(text="something")
}
\arguments{
\item{text}{the text to print (default=" ")}
}
\value{
nothing
}
\description{
\code{banner()} prints out a text in large letters much like UNIX banner
but does not truncate text at 10 characters. Instead, it will print
all the text given in a single "line".
}
\details{
Since output letters are 8 characters wide, this means, in fact, that
if you want your text to look good, you should truncate it to the
line width limit selected in your R session divided by 8.

Maybe in a future release I will modify the package to detect maximum
line width and do the truncation... for now I find it more useful not
to have explicit limits.

output
}
\examples{
banner('_______')
banner()
banner(' Hiya! ')
banner('_______')
\dontshow{ banner('Hello World!') }

}
\author{
(C) José R Valverde, CNB-SIC, 2018
}
